+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Fixed
 - Technical - Make the tests run on the non-transpiled sources.
-- Authentication - Fix an empty user id attribut in the JWT tokens.
+- Authentication - Fix an empty user id attribute in the JWT tokens.
 
 ## RELEASE 2.15.0 - 2018-09-08
 ### Added